Moved PopupMenu and encryption out of RecyclerAdapter
